src/backend/requirements.txt
# 基础依赖
fastapi<=0.90.1
uvicorn<=0.21.1
python-multipart<=0.0.6
PyYaml<=6.0.1
requests<=2.28.2
prettydiff
pycryptodome==3.19.0
pydantic